Added loading screen on superuser/ratings route
