+++ src/Components/superuser/ratings/Ratings.jsx
@@ -1,19 +1,12 @@
 import React, { useState } from 'react';
-import Select from '@material/react-select';
-import PropTypes from 'prop-types';
 import { useHistory } from 'react-router-dom';
 import { Headline4, Body2 } from '@material/react-typography';
 import Button from '@material/react-button';
+import SelectingContest from './SelectingContest';
 
-const SuperuserRatings = ({ contests }) => {
+const SuperuserRatings = () => {
   const history = useHistory();
-  const [contest, setContest] = useState(contests[0].code);
-  const onContestChange = (index, item) => setContest(item.getAttribute('data-value'));
-
-  const contestOptions = contests.map((contestOption) => ({
-    value: contestOption.code,
-    label: `${contestOption.name} (${contestOption.code})`,
-  }));
+  const [contest, setContest] = useState('None');
 
   return (
     <div className="mw7 center ma1">
@@ -21,16 +14,7 @@ const SuperuserRatings = ({ contests }) => {
       <Body2 className="mid-gray mt2 mb4 ml1">
         Ratings will be changed using Elo Ratings Algorithm
       </Body2>
-      <Select
-        className="w-100"
-        notchedOutlineClassName="pa1"
-        enhanced
-        outlined
-        label="Select Contest"
-        value={contest}
-        options={contestOptions}
-        onEnhancedChange={onContestChange}
-      />
+      <SelectingContest contest={contest} setContest={setContest} />
       <Button
         className="ma1 mt3 mb5"
         raised
@@ -40,10 +24,6 @@ const SuperuserRatings = ({ contests }) => {
       </Button>
     </div>
   );
-};
-
-SuperuserRatings.propTypes = {
-  contests: PropTypes.array.isRequired,
 };
 
 export default SuperuserRatings;
